Make transactions wait for locking to avoid deadlocks.
When the ramnesia machine detects lock it will record
locking and locked transactions and send unlock message
to the locked transaction pids when locking transactions finish.

To avoid deadlocks, a locked transaction will only wait for
transactions with higher transaction ID, if there are no
such transactions, it will restart without waiting.

Transaction restart does not create a new transaction ID
and does not rollback the transaction. It relies on lock operation
to clean locks, aquired in the previous incarnation of the transaction.