Log replayability and snapshotting.
All successfuly committed transactions are recorded to a mnesia
table together with the commit itself.
If a transaction is recorded as committed, all operations will be
skipped and return errors, cleaning up the transaction in the
machine state.
This is fine, because this commands are replays and will not return
to the client (at least should not).
Snapshots are reported on every successful commit.
Snapshots only work if there is a mnesia database saved on disk, which
is expected anyway.