Read optimisation. Read data locally and check versions on commit.
Create a versions table to store incrementing versions of each table.
On write commit, the version is updated.
Reads are performed locally, the table version is recorded in the transaction context.
On commit all read versions are checked to be up-to-date with the leader data.
If they are not - the commit needs to restart.
Transaction process will wait until leader data becomes actual on the local node.

Addresses #2

[#163603089]