Change the read optimisation.
Do not check versions on commit, because multiple unrelated
writes may update table version.
This makes concurrent updates on different keys very slow.

Instead we read the version on lock and wait for the local node
to progress to that version. Lock should take care that there is no
updates on this record. Version in this case makes sure that the local
read is up-to-date with the locker node (which should be the leader node).

More fine-grained versioning may be required for futrther optimisations.