Handle transaction process failures consistently.
Transaction processes are monitored by the locker process.
This means that transaction may be crashed from the
locker point of view, but still running.

This can cause races when a new transaction checks locks
after the down message, but commits before the "failed"
transaction commit.

This breaks the consistency guarantee.

To prevent that we may block the "failed" transaction from
committing. This can be done by locker process communicating
to the raft process and asking it to "blacklist" the transaction.

Blacklisted transactions will return an error on commit.

The locker process may clean up the "failed" transaction locks after
it gets a confirmation that the "blacklist" command was applied.
From that point the "failed" transaction cannot be committed and it's
safe to cleanup its locks.

"L" - the locker holds the transaction lock.
Other transactions cannot use the locked resource.
After DOWN message other transactions may aquire a lock on a resource already
owned by transaction 1.

Because of that there can be an inconsistency between transactions:

   tr1                     locker        machine
    |                        |              |
    |  Lock ok               |              |
    |----------------------> L1             |
    |                        L1             |
    |    DOWN                L1             |
    |----------------------> L1             |
    |                        |              |
    |         tr2            |              |
    |          |   Lock ok   |              |
    |          |------------>L2             |
    |          |             L2             |
    |          |     commit2 ok             |
    |          |--------------------------->|
    |   commit1 ok                          |
    |-------------------------------------->|

With transaction blacklisting:

   tr1                     locker        machine
    |                        |              |
    |  Lock ok               |              |
    |----------------------> L1             |
    |                        L1             |
    |    DOWN                L1             |
    |----------------------> L1             |
    |                        L1             |
    |                        L1 blacklist   |
    |         tr2            L1-----------> |
    |          |             L1             |
    |          | Lock wait   L1             |
    |          | ----------> L1 blacklisted |
    |          |             L1 <---------- |
    |          | retry lock  |              |
    |          |-----------> L2             |
    |          |             L2             |
    |          |             L2             |
    |          |     commit2 ok             |
    |          |--------------------------->|
    |   commit1 rejected (blacklisted)      |
    |-------------------------------------->|