Support inefficient QLC API.
Select with limit does not require strict number of records in response.
Limit is a recommendation according to mnesia:select/4 docs.
This means QLC can work even with a simplistic select/4, which will return
all matching records (like select/2) and '$end_of_table' as a continuation.
This allows the QLC api to be used, but it's not the most efficient solution.